### 2024-03-30 #6559 @editorialbot
[PRE REVIEW]: VineCopulas: an open-source Python package for vine copula modelling
**Submitting author:** @judithclaassen (<a href="http://orcid.org/0009-0000-2979-1297">Judith Claassen</a>)
**Repository:** https://github.com/VU-IVM/VineCopulas
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.5
