+++ paper/paper.bib
@@ -115,7 +115,7 @@ year = {2021}
   publisher = "Chapman and Hall/CRC",
   month     =  may,
   year      =  1997,
-  doi       = "https://doi.org/10.1201/9780367803896"
+  doi       = {https://doi.org/10.1201/9780367803896}
 }
 
 @Article{Eilander,
