+++ paper/paper.bib
@@ -15,7 +15,7 @@ year = {2022}
 @book{Joe2014,
 address = {New York},
 author = {Joe, Harry},
-doi = {https://doi.org/10.1201/b17116},
+doi = {10.1201/b17116},
 edition = {1st},
 isbn = {9780429103186},
 publisher = {Chapman and Hall/CRC},
