### 2024-05-02 #6559 @judithclaassen
@ymzayek , Thank you so much for agreeing to be the editor for our paper. We would like to suggest two reviewers that are not on the list of people that have already agreed to review JOSS, however, their expertise are very relevant to this package:

- Anais Couasnon (couasnonanais)
- Zengchao Hao (haozcbnu)

As for the missing and invalid DOI's, I addressed one invalid DOI and added one missing DOI. the other 'missing' DOI's are sources that do not have a DOI, such as other packages.

Thank you very much in advance
