### 2024-05-06 #6559 @ymzayek
Hello @couasnonanais and @haozcbnu,

Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software)?

VineCopulas: an open-source Python package for vine copula modelling

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/6559).

Thank you in advance!
