### 2024-05-07 #6559 @couasnonanais
Hi,

I am also happy to review it.

Anaïs

Le lun. 6 mai 2024 à 09:36, Yasmin ***@***.***> a écrit :

> Hello @couasnonanais <https://github.com/couasnonanais> and @haozcbnu
> <https://github.com/haozcbnu>,
>
> Would you be willing to assist in reviewing this submission for JOSS
> (Journal of Open Source Software)?
>
> JOSS publishes articles about open source research software. The
> submission I'd like you to review is titled:
>
> VineCopulas: an open-source Python package for vine copula modelling
>
> You can find more information at the top of this Github issue (#6559
> <https://github.com/openjournals/joss-reviews/issues/6559>).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged. If you have
> any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will
> start the review process in a separate thread. In that thread, there will
> be 23 check items for each single reviewer.
>
> Thank you in advance!
>
