### 2024-05-07 #6728 @editorialbot
[REVIEW]: VineCopulas: an open-source Python package for vine copula modelling
**Submitting author:** @judithclaassen (<a href="http://orcid.org/0009-0000-2979-1297">Judith Claassen</a>)
**Repository:** https://github.com/VU-IVM/VineCopulas
**Branch with paper.md** (empty if default branch):
**Version:** 2.0.0
