### 2024-05-07 #6728 @ymzayek
@haozcbnu and @couasnonanais

Dear reviewers, you can start your review by creating your tasklist with the following command:

`@editorialbot generate my checklist`

In that list, there are several tasks. Whenever you perform a task, you can check on the corresponding checkbox. You can also reference the JOSS reviewer guidelines which is linked in first comment in this thread. Since the review process of JOSS is interactive, you can always interact with the author, the other reviewers, and the editor during the process. You can open issues and pull requests in the target repo. Please mention the url of this page in there so that we can keep tracking what is going on.

Thank you in advance.
