### 2024-05-29 #6728 @judithclaassen
> > Here are several comments/suggestions.
> > Lines 9-10: Here some references can be given to introduce the basic concept of Copula.
> > Lines 15-16: please check the punctuation mark.
> > Lines 19-22: Here the concept of the "curse of dimensionality " can be introduced to demonstrate the challenge in multivariate dependence modeling.
> > Captions of Figure 1: D) What do you mean by " As not every vine copula structure is suitable to generate conditional samples of every variable"? E) Drawing conditional samples of the variable 1 given specific values of variables 2 and 3? This needs to be clear.
> > Lines 59-61: To emphasize the usefulness of this package, the authors can highlight the large number of variables or hazards (>2) in the multi-hazard phenomenon.
>
> @haozcbnu can you clarify the status regarding these comments? Any updates there?

@haozcbnu Thank you so much for providing us with your review. @ymzayek I would prefer to incorporate the suggestions after having received @couasnonanais review as well, if possible.
