### 2024-06-25 VU-IVM/VineCopulas#15 @couasnonanais
JOSS review : Software paper
Dear authors of the VineCopulas package and @judithclaassen ,

Apologies for the delay in my review. I very much enjoyed having a look at the package and I think it is a great contribution to the Python community.  You will find below some comments concerning the paper on top of the ones mentioned by @haozcbnu . I will be opening a separate issue for the documentation.

Concerning Figure 1: I think that this figure summarizes very well the main functionalities of the package and therefore I would suggest to clarify the figure.
- [x] It is not obvious at first that the letters in the caption refer to the arrows.
- [x] In the Table with the Data, I would suggest to change the notation to be in line with common mathematical notations. Therefore, I would add next to Variable 1 ($X_1$), etc. In the table, the notation then becomes $x^1_1$ until $x^n_1$ and for $X_2$: $x^1_2$ until $x^n_2$. Calling this table as Table 1 data or somehting like that can then clarify the caption (see next point)
- [x] In the caption after (A), you could then mention that "Samples from Table 1 - data, consisting of both continuous and discrete varibles (plotted in blue) are transformed into pseudo-observations using their marginal distributions (shown in green)". It would be good to mention pseudo-observations/pseudo-data there too (I provide a suggestion but feel free to use it or not)

Concerning the Statement of need:
- [x] It is not clear who the target audience is. I understand after reading the text that one type of audience is Python users and from the examples, someone who has some mathematical background. You could add a few sentences about this in the paper to explicitely adress this (since this is in the review checklist)
- [x] I understand that the paragraph between line 49 and 53 lists the unique features of the package. I would make this more explicit. (for e.g. "The unique features of this package is that is integrates....") or even using a bullet point list. Here I would clarify "fitting CDF and PDF" given the question you received as an issue (issue #14 )
- [ ] You could then extend this bullet list to the main functionalities of the package (as highlighted in the examples)

Concerning references:
- [x] I would add all references that are mentioned in the example Python notebooks as I would argue that they are key references.
- [ ] In the Statement of need, I would highlight the paragraph

Typos etc:
- [x] line 50: CDF --> you mean cumulative here, correct?
- [x] line 51: random sample generation. (sampling --> sample)
- [x] line 58: Python
- [x] line 54: reference missing for the copulas package?
- [x] The last sentences reads a bit strange because these studies have used alternative coding languages or Python package. Maybe rephrase to put the emphasis on the need for more complete package to apply for these high dimensionality and complex problem. But that also can welcome contributions from the growing Python community.
