### 2024-06-25 VU-IVM/VineCopulas#16 @couasnonanais
JOSS review : Documentation
Dear authors of the VineCopulas package and @judithclaassen ,

Apologies for the delay in my review. I very much enjoyed having a look at the package and I think it is a great contribution to the Python community. You will find below a few comments concerning the Documentation. Especially users that are not very familiar with copulas or VineCopulas may benefit from these changes

For example, to help users,  you could consider:
- [x] Adding the Getting Started section  to the home page of the Documentation
- [x] Maybe in general expand the introduction to the homepage.
- [x] Point towards examples after each bullet point of what the package can do
