### 2024-06-25 VU-IVM/VineCopulas#15 @couasnonanais
I forgot to mention that it would be good to mention or point the user towards the list of univariate distributions available for the fit of the marginal distributions. And explicitely mention the scipy package.
