### 2024-06-25 #6728 @couasnonanais
Hi @ymzayek, @haozcbnu and @judithclaassen,

Apologies for the delay in my review. I very much enjoyed having a look at the package and I think it is a great contribution to the Python community.

I have added my comments as issues (https://github.com/VU-IVM/VineCopulas/issues/16#issue-2373323282 and https://github.com/VU-IVM/VineCopulas/issues/15#issue-2372600177).

Thank you,

Anaïs
