+++ paper/paper.md
@@ -31,29 +31,30 @@ bibliography: paper.bib
 ---
 
 # Summary
-A copula method can be used to describe the dependency structure between several random variables. Copula-methods are used widely in various research fields across different disciplines, ranging from  finance to the bio-geophysical sciences. While some other multivariate distributions, for instance a multivariate normal distribution, allow for a highly symmetric dependency structure with the same univariate and multivariate marginal distributions, copulas can model the joint distribution of multiple random variables separately from their marginal distribution [@Czado2022].
+A copula method can be used to describe the dependency structure between several random variables. Copula-methods are used widely in various research fields across different disciplines, ranging from  finance to the bio-geophysical sciences [@dimann_2013_selecting; @klein_2020_bayesian; @Mitskopoulos2022]. While some other multivariate distributions, for instance a multivariate normal distribution, allow for a highly symmetric dependency structure with the same univariate and multivariate marginal distributions, copulas can model the joint distribution of multiple random variables separately from their marginal distribution [@sklar1959fonctions; @Czado2022].
 
-Once a copula distribution has been modelled. They allow for random samples of the data to be generated, as well as conditional samples. For example, if a copula has been fit between people's height and weight, this copula can create random correlated samples of both variables as well as conditional samples, e.g., samples of weight given a specific height.
+Once a copula distribution has been modelled, they allow for random samples of the data to be generated, as well as conditional samples. For example, if a copula has been fit between people's height and weight, this copula can create random correlated samples of both variables as well as conditional samples, e.g., samples of weight given a specific height.
 
-Although copulas are an excellent tool to model dependencies in bivariate data, data with two variables, there are only a limited number of copulas capable of modelling larger multivariate datasets, for example, the Gaussian and Student-t copula. However, when modelling the dependencies between a large number of different variables, a more flexible multivariate modelling tool may be required that does not assume a single copula to capture all the individual dependencies. To this end, vine copulas have been proposed as a method to construct a multivariate model with the use of bivariate copulas as building blocks [@Joe1997; @cooke2001; @bedford]. 
+Although copulas are an excellent tool to model dependencies in bivariate data, data with two variables, there are only a limited number of copulas capable of modelling larger multivariate datasets, for example, the Gaussian and Student-t copula. However, when modelling the dependencies between a large number of different variables, a more flexible multivariate modelling tool may be required that does not assume a single copula to capture all the individual dependencies. To this end, vine copulas have been proposed as a method to construct a multivariate model with the use of bivariate copulas as building blocks [@Joe1997; @cooke2001; @bedford; @aas_2009_paircopula]. 
 
-In the previous example related to height and weight, a vine copula could be used to also model age in relation to height and weight. Like bivariate copulas, vine copulas allow the user to generate random and conditional samples. However, to draw conditional samples from a vine copula for a specific variable, the vine copula has to be structured in such a way that the order in which the samples are generated draws the variable of interest last, i.e. the sample is conditioned on the preceding samples of other variables. For example, if one wants to generate a conditional sample of height, the samples of age and weight have to be provided first.  Additionally, while it is more common to use copulas for continuous data, such as weight and height, methods have been developed to also allow for discrete data, such as age, to be modelled [@Mitskopoulos2022]. 
[…]
