+++ paper/paper.bib
@@ -76,7 +76,7 @@ year = {2021}
 }
 
 @misc{inc_copulas,
-  author = {Inc, DataCebo},
+  author = {DataCebo},
   title = {copulas: Create tabular synthetic data using copulas-based modeling.},
   url = {https://pypi.org/project/copulas/},
   urldate = {2024-03-07},
@@ -142,3 +142,78 @@ PAGES = {2701--2723},
 URL = {https://hess.copernicus.org/articles/21/2701/2017/},
 DOI = {10.5194/hess-21-2701-2017}
 }
+
+
+@inbook{Morales,
+title = "Counting Vines",
+keywords = "Boekdeel internat.wet",
+author = "{Morales Napoles}, O",
+year = "2011",
+language = "English",
+isbn = "978-981-4299-87-9",
+pages = "189--218",
+editor = "D Kurowicka and H Joe",
+booktitle = "Dependence Modeling. Vine Copula Handbook",
+publisher = "World Scientific Publishing",
+address = "Singapore",
+}
+
+
+@article{sklar1959fonctions,
+  title={Fonctions de repartition an dimensions et leurs marges},
+  author={Sklar, M},
+  journal={Publ. inst. statist. univ. Paris},
+  volume={8},
+  pages={229--231},
+  year={1959}
+}
+
+
+@article{aas_2009_paircopula,
+  author = {Aas, Kjersti and Czado, Claudia and Frigessi, Arnoldo and Bakken, Henrik},
+  month = {04},
+  pages = {182-198},
+  title = {Pair-copula constructions of multiple dependence},
+  doi = {10.1016/j.insmatheco.2007.02.001},
+  urldate = {2020-11-13},
+  volume = {44},
+  year = {2009},
+  journal = {Insurance: Mathematics and Economics}
+}
+
+@article{cooke_2015_sampling,
+  author = {Cooke, R.M. and Kurowicka, D. and Wilson, K.},
+  month = {06},
+  pages = {4-18},
+  title = {Sampling, conditionalizing, counting, merging, searching regular vines},
+  doi = {10.1016/j.jmva.2015.02.001},
+  urldate = {2022-09-24},
+  volume = {138},
+  year = {2015},
+  journal = {Journal of Multivariate Analysis}
+}
+
+@article{dimann_2013_selecting,
+  author = {Dißmann, J. and Brechmann, E.C. and Czado, C. and Kurowicka, D.},
+  month = {03},
+  pages = {52-69},
+  title = {Selecting and estimating regular vine copulae and application to financial returns},
+  doi = {10.1016/j.csda.2012.08.010},
+  urldate = {2022-03-26},
+  volume = {59},
+  year = {2013},
+  journal = {Computational Statistics & Data Analysis}
+}
+
+@article{klein_2020_bayesian,
+  author = {Klein, Nadja and Kneib, Thomas and Marra, Giampiero and Radice, Rosalba},
+  month = {01},
+  pages = {121-152},
+  publisher = {Elsevier BV},
+  title = {Bayesian mixed binary-continuous copula regression with an application to childhood undernutrition},
+  doi = {10.1016/b978-0-12-815862-3.00011-1},
+  urldate = {2024-07-05},
+  year = {2020},
+  journal = {Elsevier eBooks}
+}
+
