+++ doc/index.rst
@@ -3,10 +3,10 @@ Welcome to VineCopulas's documentation!
 
 `VineCopulas` is a Python package that is able to:
 
-* Fit both bivariate and vine copulas
-* Simulate from both bivariate and vine copulas
-* Allow for both discrete as well as continuous input data
-* Draw conditional samples for any variables of interest with the use of bivariate copulas and different vine structures
+* Fit both `bivariate <https://vinecopulas.readthedocs.io/en/latest/bivariatecopulas.html#Fitting-a-Copula>`_ and `vine copulas <https://vinecopulas.readthedocs.io/en/latest/vinecopulas.html#Fitting-a-Vine-Copula>`_
+* Simulate from both `bivariate <https://vinecopulas.readthedocs.io/en/latest/bivariatecopulas.html#Generating-Random-Samples-from-a-Copula>`_ and `vine copulas <https://vinecopulas.readthedocs.io/en/latest/vinecopulas.html#Generating-Random-Samples-from-a-Vine-Copula>`_
+* Allow for both `discrete <https://vinecopulas.readthedocs.io/en/latest/bivariatecopulas.html#Discrete-Variables>`_ as well as  `continuous <https://vinecopulas.readthedocs.io/en/latest/bivariatecopulas.html#Probability-Integral-Transform-with-Fitted-CDF>`_ input data
+* Draw conditional samples for any variables of interest with the use of `bivariate copulas <https://vinecopulas.readthedocs.io/en/latest/bivariatecopulas.html#Generating-Conditional-Random-Samples-from-a-Copula>`_ and different `vine <https://vinecopulas.readthedocs.io/en/latest/vinecopulas.html#Conditional-Sampling>`_ structures
 
 Installation
 -------
