+++ paper/paper.md
@@ -42,15 +42,15 @@ In the previous example related to height and weight, a vine copula could be use
 `VineCopulas` is a Python package that is able to fit and simulate both bivariate and vine copulas. This package allows for both discrete as well as continuous input data, and can draw conditional samples for any variables of interest with the use of different vine structures (see Figure 1).
 
 
-![A schematic representation of VineCopulas functionalities, where the lettering refers to the different arrows (functions). A) Samples from data, consisting of both continuous and discrete variables, are transformed using their marginal distributions into pseudo data. B) A vine copula is fit to the transformed data. Here, the first tree has nodes containing the variables and edges denoting the bivariate dependencies. The edges in the second tree denote the dependency between all variables. C) Using the fitted vine copula, random samples are generated. D) As not every vine copula structure is suitable to generate conditional samples of every variable, due to its inherent sampling order, a vine copula can also be fit conditionally. Here, a vine copula is fit conditionally for variable 1. E) The conditionally fit vine copula is used to draw conditional samples of variable 1 given specific values of variables 2 and 3.\label{fig:schematic}](figure1.png)
+![A schematic representation of VineCopulas functionalities, where the lettering refers to the different arrows (Python functions). A) Samples from Table 1 - data, consisting of both continuous and discrete variables (plotted in blue) are transformed into pseudo-observations using their marginal distributions (shown in green). B) A vine copula is fit to the transformed data. Here, the first tree has nodes containing the variables and edges denoting the bivariate dependencies. The edges in the second tree denote the dependency between all variables. C) Using the fitted vine copula, random samples are generated. D) As not every vine copula structure is suitable to generate conditional samples of every variable, due to its inherent sampling order, a vine copula can also be fit conditionally. Here, a vine copula is fit conditionally for variable 1. E) The conditionally fit vine copula is used to draw conditional samples of variable 1 given specific values of variables 2 and 3.\label{fig:schematic}](figure1.png)
 
 # Statement of need
 
 The programming language R is widely known as the most advanced statistical programming language and hence has many well-developed packages for copulas, such as `copula` [@copulapackage],  `VineCopula` [@vinecopulapackage], and `CDVineCopulaConditional` [@Bevacqua]. However, with the open source programming language Python gaining more popularity for statistical programming, there is an increasing interest in Python-based copula packages. Therefore, we have developed the package `VineCopulas`, a pure Python implementation for (vine) copulas. 
 
-`VineCopulas` integrates many of the standard copula package features, including fitting, cumulative distribution functions (CDFs) of bivariate copulas, Probability Density Function (PDFs), and random sample generation. In addition, the package also enables the user to generate conditional samples, fit vine structures to facilitate specific conditional probabilities and fit as well as simulate discrete data, all of which are unique to have in a single package.
+`VineCopulas` integrates many of the standard copula package features, including fitting, Probability Density Function (PDFs), and random sample generation for bivariate and vine copulas. This package can also fit the best marginal distributions of the individual variables based on the univariate distributions available in the statistical Python package `SciPy` [@2020SciPy-NMeth]. Furthermore, the `VineCopulas` can compute cumulative distribution functions (CDFs) of bivariate copulas. In addition, the package also enables the user to generate conditional samples, fit vine structures to facilitate specific conditional probabilities and fit as well as simulate discrete data, all of which are unique to have in a single package.
 
[…]
