### 2024-07-08 8317c87
new paper version and changes to documentation
