+++ paper/paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-A copula method can be used to describe the dependency structure between several random variables. Copula-methods are used widely in various research fields across different disciplines, ranging from  finance to the bio-geophysical sciences [@dimann_2013_selecting; @klein_2020_bayesian; @Mitskopoulos2022]. While some other multivariate distributions, for instance a multivariate normal distribution, allow for a highly symmetric dependency structure with the same univariate and multivariate marginal distributions, copulas can model the joint distribution of multiple random variables separately from their marginal distribution [@sklar1959fonctions; @Czado2022].
+A copula method can be used to describe the dependency structure between several random variables. Copula methods are used widely in various research fields across different disciplines, ranging from  finance to the bio-geophysical sciences [@dimann_2013_selecting; @klein_2020_bayesian; @Mitskopoulos2022]. While some other multivariate distributions, for instance a multivariate normal distribution, allow for a highly symmetric dependency structure with the same univariate and multivariate marginal distributions, copulas can model the joint distribution of multiple random variables separately from their marginal distribution [@sklar1959fonctions; @Czado2022].
 
 Once a copula distribution has been modelled, they allow for random samples of the data to be generated, as well as conditional samples. For example, if a copula has been fit between people's height and weight, this copula can create random correlated samples of both variables as well as conditional samples, e.g., samples of weight given a specific height.
 
@@ -53,7 +53,7 @@ The programming language R is widely known as the most advanced statistical prog
 While there are two well-used python copula packages, `copulas` [@inc_copulas], and `pyvinecopulib` [@pyvinecoplib], neither of these packages includes the above-mentioned unique features. Furthermore, `copulas` is mostly suitable for bivariate copulas, and has limited vine copula capabilities, while `pyvinecopulib` is a C++ library with a Python interface, meaning that it is not fully Python-based, and therefore less adaptable for a Python user. Therefore, `VineCopulas` is targeted towards data analysts, researchers and modellers in various fields, who are python users or require functionality specifically for discrete data and conditional sampling.
 
 
-` VineCopulas` is currently being used in a study on multi-hazards to model the dependencies between different natural hazard intensities. For this study, the ability to generate conditional samples is required to evaluate possible magnitudes of one natural hazard given multiple others e.g., levels of extreme precipitation given specific extreme wind speeds and relative humidity. The capability to also simulate discrete data may be useful for hazards with intensity measures of a discrete nature, such as the Volcanic Explosivity Index (VEI). Applications of this type, are growing in the field of compound and multi-hazard risk research [@Eilander; @Bevacqua_2]. ` VineCopulas` will allow python users to continue this research at a higher dimensionality,  showing the clear need for this package.
+` VineCopulas` is currently being used in a study on multi-hazards to model the dependencies between different natural hazard intensities. For this study, the ability to generate conditional samples is required to evaluate possible magnitudes of one natural hazard given multiple others e.g., levels of extreme precipitation given specific extreme wind speeds and relative humidity. The capability to also simulate discrete data may be useful for hazards with intensity measures of a discrete nature, such as the Volcanic Explosivity Index (VEI). Applications of this type, are growing in the field of compound and multi-hazard risk research [@Eilander; @Bevacqua_2].` VineCopulas` will allow python users to continue this research at a higher dimensionality, showing the clear need for this package.
 
 
 # Acknowledgements
