+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "VineCopulas"
-version = "0.2.5"
+version = "2.0.0"
 authors = [
     {name = "Judith Claassen", email = "j.n.claassen@vu.nl"},
 ]
