### 2024-07-08 #6728 @judithclaassen
@ymzayek Thank you for checking on the review process. @couasnonanais @haozcbnu thank you both so much for taking the time to review. Your comments have helped us to improve the manuscript and documentation. I would like to highlight how the comments have been implemented.

**Response to @haozcbnu :**

**Lines 9-10: Here some references can be given to introduce the basic concept of Copula.**

Reference added in L9-10 and L15

**Lines 15-16: please check the punctuation mark.**

corrected

**Lines 19-22: Here the concept of the "curse of dimensionality " can be introduced to demonstrate the challenge in multivariate dependence modeling.**

While this can be relevant to, for example defining the CDF of a vine copula, we think it is slightly to specific for a JOSS paper, which is focused more on the functionality and need for a broader audience. We do have some information about the increasing complexity of higher dimensions in the documentation.

**Captions of Figure 1: D) What do you mean by " As not every vine copula structure is suitable to generate conditional samples of every variable"? E) Drawing conditional samples of the variable 1 given specific values of variables 2 and 3? This needs to be clear.**

Added clarification in the caption

**Lines 59-61: To emphasize the usefulness of this package, the authors can highlight the large number of variables or hazards (>2) in the multi-hazard phenomenon.**

added an additional variable in L 68-69

**Response to @couasnonanais

Paper:

 It is not obvious at first that the letters in the caption refer to the arrows.**

Clarified in the caption

 **In the Table with the Data, I would suggest to change the notation to be in line with common mathematical notations. Therefore, I would add next to Variable 1 (𝑋1), etc. In the table, the notation then becomes 𝑥11 until 𝑥1𝑛 and for 𝑋2: 𝑥21 until 𝑥2𝑛. Calling this table as Table 1 data or somehting like that can then clarify the caption (see next point)**

Table has been adjusted

 **In the caption after (A), you could then mention that "Samples from Table 1 - data, consisting of both continuous and discrete varibles (plotted in blue) are transformed into pseudo-observations using their marginal distributions (shown in green)". It would be good to mention pseudo-observations/pseudo-data there too (I provide a suggestion but feel free to use it or not)**

Suggestion implemented

 **It is not clear who the target audience is. I understand after reading the text that one type of audience is Python users and from the examples, someone who has some mathematical background. You could add a few sentences about this in the paper to explicitely adress this (since this is in the review checklist)**

Addressed in L63-65

**I understand that the paragraph between line 49 and 53 lists the unique features of the package. I would make this more explicit. (for e.g. "The unique features of this package is that is integrates....") or even using a bullet point list. Here I would clarify "fitting CDF and PDF" given the question you received as an issue (issue PDF and CDF calculation! #14 )**

Adjusted L 50-52 to incorporate this. Also added a Density function to the package and will address the issue.

**You could then extend this bullet list to the main functionalities of the package (as highlighted in the examples)**

As there is a bullet point list on the documentation home page, github and PyPI homepage, we have decided not to implement a bullet point list in the paper but keep the functionality in a paragraph instead.

 **I would add all references that are mentioned in the example Python notebooks as I would argue that they are key references.**

Implemented across the paper

 **In the Statement of need, I would highlight the paragraph**

Unsure which paragraph to highlight. Please clarify if relevant

**Typos etc:
•	 line 50: CDF --> you mean cumulative here, correct?
•	 line 51: random sample generation. (sampling --> sample)
•	 line 58: Python
•	 line 54: reference missing for the copulas package?
•	 The last sentences reads a bit strange because these studies have used alternative coding languages or Python package. Maybe rephrase to put the emphasis on the need for more complete package to apply for these high dimensionality and complex problem. But that also can welcome contributions from the growing Python community**

All typos have been addressed

**it would be good to mention or point the user towards the list of univariate distributions available for the fit of the marginal distributions. And explicitely mention the scipy package.**

Added in L52-53

**Adding the Getting Started section to the home page of the Documentation**

Added section to the home page of the documentation

**Maybe in general expand the introduction to the homepage.**

Addressed by implementing the previous comment and the following comment that already make the Homepage of the documentation more complete.

**Point towards examples after each bullet point of what the package can do**

in the bullet point list on the homepage we hade added hyperlinks to the relevant information in the documentation to guide the reader.
