### 2024-07-10 #6728 @couasnonanais
Hi @judithclaassen,

Most of my points have been addressed, thanks. The scientific notation **in** Table 1:Data has not been adjusted however. I think this is worth correcting, otherwise it is quite confusing.
