### 2024-07-10 #6728 @judithclaassen
> Hi @judithclaassen,
>
> Most of my points have been addressed, thanks. The scientific notation **in** Table 1:Data has not been adjusted however. I think this is worth correcting, otherwise it is quite confusing.

@couasnonanais thank you for checking the revisions. Apologies, I overlooked that part of the comment. I have made the recommended changes to figure 1.
