### 2024-07-15 #6728 @ymzayek
Thank you @couasnonanais and @judithclaassen for the updates.

@couasnonanais can you have a look at the [checklist above](https://github.com/openjournals/joss-reviews/issues/6728#issuecomment-2114932010) and update it? Is there is anything left unaddressed from your comments?
