### 2024-07-24 #6728 @ymzayek
@judithclaassen can you recheck the missing DOI's in https://github.com/openjournals/joss-reviews/issues/6728#issuecomment-2247417188 and correct them and justify each missing one

Can you also have a look at the additional author tasks in https://github.com/openjournals/joss-reviews/issues/6728#issuecomment-2247421142 and complete them?
