### 2024-07-29 #6728 @judithclaassen
@ymzayek I just double checked the missing DOI's and everything is correct. The first missing DOI's are of other software packages that do not have a DOI and the last missing DOI's are book chapters etc. which also do not have a DOI.

I have also completed the authors tasks:

- I can confirm that the authors affiliations and ORCIDs are correct.
- I have made the release with the changes from the review. this release is 2.0.0 (https://github.com/VU-IVM/VineCopulas/tree/v2.0.0)
- I have archived the release on Zenodo with the DOI [10.5281/zenodo.13121560](https://doi.org/10.5281/zenodo.13121560)
- author list and ORCID's match those in the manuscript
- license listed is the same both on Github and Zenodo
