### 2024-09-02 #6728 @judithclaassen
> @judithclaassen the title in the zenodo archive doesn't match the paper's. It has to match exactly. Can you update that? Also in the Zenodo archive the last author shows upas a third affiliation but I think as I understand from what is shown on the paper list of authors, they are from the same institution listed as `1`

@ymzayek Thank you for pointing this out. The title and affiliation have now been adjusted.
