+++ paper/paper.md
@@ -50,10 +50,10 @@ The programming language R is widely known as the most advanced statistical prog
 
 `VineCopulas` integrates many of the standard copula package features, including fitting, Probability Density Function (PDFs), and random sample generation for bivariate and vine copulas. This package can also fit the best marginal distributions of the individual variables based on the univariate distributions available in the statistical Python package `SciPy` [@2020SciPy-NMeth]. Furthermore, the `VineCopulas` can compute cumulative distribution functions (CDFs) of bivariate copulas. In addition, the package also enables the user to generate conditional samples, fit vine structures to facilitate specific conditional probabilities and fit as well as simulate discrete data, all of which are unique to have in a single package.
 
-While there are two well-used python copula packages, `copulas` [@inc_copulas], and `pyvinecopulib` [@pyvinecoplib], neither of these packages includes the above-mentioned unique features. Furthermore, `copulas` is mostly suitable for bivariate copulas, and has limited vine copula capabilities, while `pyvinecopulib` is a C++ library with a Python interface, meaning that it is not fully Python-based, and therefore less adaptable for a Python user. Therefore, `VineCopulas` is targeted towards data analysts, researchers and modellers in various fields, who are python users or require functionality specifically for discrete data and conditional sampling.
+While there are two well-used Python copula packages, `copulas` [@inc_copulas], and `pyvinecopulib` [@pyvinecoplib], neither of these packages includes the above-mentioned unique features. Furthermore, `copulas` is mostly suitable for bivariate copulas, and has limited vine copula capabilities, while `pyvinecopulib` is a C++ library with a Python interface, meaning that it is not fully Python-based, and therefore less adaptable for a Python user. Therefore, `VineCopulas` is targeted towards data analysts, researchers and modellers in various fields, who are Python users or require functionality specifically for discrete data and conditional sampling.
 
 
-`VineCopulas` is currently being used in a study on multi-hazards to model the dependencies between different natural hazard intensities. For this study, the ability to generate conditional samples is required to evaluate possible magnitudes of one natural hazard given multiple others e.g., levels of extreme precipitation given specific extreme wind speeds and relative humidity. The capability to also simulate discrete data may be useful for hazards with intensity measures of a discrete nature, such as the Volcanic Explosivity Index (VEI). Applications of this type, are growing in the field of compound and multi-hazard risk research [@Eilander; @Bevacqua_2].` VineCopulas` will allow python users to continue this research at a higher dimensionality, showing the clear need for this package.
+`VineCopulas` is currently being used in a study on multi-hazards to model the dependencies between different natural hazard intensities. For this study, the ability to generate conditional samples is required to evaluate possible magnitudes of one natural hazard given multiple others e.g., levels of extreme precipitation given specific extreme wind speeds and relative humidity. The capability to also simulate discrete data may be useful for hazards with intensity measures of a discrete nature, such as the Volcanic Explosivity Index (VEI). Applications of this type, are growing in the field of compound and multi-hazard risk research [@Eilander; @Bevacqua_2]. ` VineCopulas` will allow Python users to continue this research at a higher dimensionality, showing the clear need for this package.
 
 
 # Acknowledgements
