+++ paper/paper.md
@@ -53,7 +53,7 @@ The programming language R is widely known as the most advanced statistical prog
 While there are two well-used Python copula packages, `copulas` [@inc_copulas], and `pyvinecopulib` [@pyvinecoplib], neither of these packages includes the above-mentioned unique features. Furthermore, `copulas` is mostly suitable for bivariate copulas, and has limited vine copula capabilities, while `pyvinecopulib` is a C++ library with a Python interface, meaning that it is not fully Python-based, and therefore less adaptable for a Python user. Therefore, `VineCopulas` is targeted towards data analysts, researchers and modellers in various fields, who are Python users or require functionality specifically for discrete data and conditional sampling.
 
 
-`VineCopulas` is currently being used in a study on multi-hazards to model the dependencies between different natural hazard intensities. For this study, the ability to generate conditional samples is required to evaluate possible magnitudes of one natural hazard given multiple others e.g., levels of extreme precipitation given specific extreme wind speeds and relative humidity. The capability to also simulate discrete data may be useful for hazards with intensity measures of a discrete nature, such as the Volcanic Explosivity Index (VEI). Applications of this type, are growing in the field of compound and multi-hazard risk research [@Eilander; @Bevacqua_2]. ` VineCopulas` will allow Python users to continue this research at a higher dimensionality, showing the clear need for this package.
+`VineCopulas` is currently being used in a study on multi-hazards to model the dependencies between different natural hazard intensities. For this study, the ability to generate conditional samples is required to evaluate possible magnitudes of one natural hazard given multiple others e.g., levels of extreme precipitation given specific extreme wind speeds and relative humidity. The capability to also simulate discrete data may be useful for hazards with intensity measures of a discrete nature, such as the Volcanic Explosivity Index (VEI). Applications of this type, are growing in the field of compound and multi-hazard risk research [@Eilander; @Bevacqua_2]. `VineCopulas` will allow Python users to continue this research at a higher dimensionality, showing the clear need for this package.
 
 
 # Acknowledgements
