### 2024-09-11 3cc689a
Update paper.bib
