+++ paper/paper.bib
@@ -85,7 +85,7 @@ year = {2021}
 
 
 @conference{cooke2001,
-  title = "Monte Carlo simulation of vine dependent random variables for applications in uncertainty analysis",
+  title = "Monte {C}arlo simulation of vine dependent random variables for applications in uncertainty analysis",
   abstract = "This paper examines Monte Carlo simulation of vine dependent random variables for applications in uncertainty analysis.",
   keywords = "monte carlo method, simulation, vine dependent, analysis",
   author = "T.J. Bedford and R. Cooke",
@@ -120,7 +120,7 @@ year = {2021}
 
 @Article{Eilander,
 AUTHOR = {Eilander, D. and Couasnon, A. and Sperna Weiland, F. C. and Ligtvoet, W. and Bouwman, A. and Winsemius, H. C. and Ward, P. J.},
-TITLE = {Modeling compound flood risk and risk reduction using a globally applicable framework: a pilot in the Sofala province of Mozambique},
+TITLE = {Modeling compound flood risk and risk reduction using a globally applicable framework: a pilot in the {S}ofala province of {M}ozambique},
 JOURNAL = {Natural Hazards and Earth System Sciences},
 VOLUME = {23},
 YEAR = {2023},
@@ -133,7 +133,7 @@ DOI = {10.5194/nhess-23-2251-2023}
 
 @Article{Bevacqua_2,
 AUTHOR = {Bevacqua, E. and Maraun, D. and Hob{\ae}k Haff, I. and Widmann, M. and Vrac, M.},
-TITLE = {Multivariate statistical modelling of compound events via pair-copula constructions: analysis of floods in Ravenna (Italy)},
+TITLE = {Multivariate statistical modelling of compound events via pair-copula constructions: analysis of floods in {R}avenna ({I}taly)},
 JOURNAL = {Hydrology and Earth System Sciences},
 VOLUME = {21},
 YEAR = {2017},
