### 2024-09-11 #6728 @judithclaassen
> 👋 @judithclaassen - I just need you to fix the following in the paper before I can move to accept this for publication:
>
> * [ ]  LINE 58:  the "p" in "python" should be capitalized
> * [ ]  LINE 64:  the "p" in "python" should be capitalized
> * [ ]  LINE 73:  there is a missing space between "...2023)." and "VineCopulas..."
> * [ ]  LINE 73:  the "p" in "python" should be capitalized
> * [ ]  Appropriate capitalization should be maintained in the references.  For example, in LINE 85 the "c" in "carlo" should be capitalized.  This can be done in the bib file using curly brackets around characters you wish to maintain the formatting of.  There are others like this throughout your references, so please review and correct these.
>
> Let me know when you have addressed these items. Thanks!

@crvernon thank you for doing a final check of the paper. I have addressed the items you listed. let me know if there is anything else I can do.
